first commit
Lots of bug fixes needed
