+++ .vscode/settings.json
@@ -1,3 +1,2 @@
 {
-    "julia.environmentPath": "/Users/kishorehari/Desktop/Boolean.jl"
 }
