+++ src/Boolean.jl
@@ -1,0 +1,9 @@
+module Boolean
+
+
+# Write your package code here.
+include("bmodel.jl")
+# end Boolean
+
+# include("bmodel.jl")
+# export bmodel_reps
