+++ src/dependencies.jl
@@ -1,4 +1,4 @@
-using Pkg
+# using Pkg
 using CSV
 using DataFrames
 using SparseArrays
