+++ test/runtests.jl
@@ -1,0 +1,6 @@
+using Boolean
+using Test
+
+@testset "Boolean.jl" begin
+    # Write your tests here.
+end
